fixed inner classes
